refactor: api 폴더명수정
